requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
openai>=1.0.0
python-dotenv>=0.19.0
textstat>=0.7.0
nltk>=3.8
selenium>=4.15.0
Pillow>=10.0.0
matplotlib>=3.7.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
wordcloud>=1.9.0
pyspellchecker>=0.7.0
readability>=0.3.1
langdetect>=1.0.9
whois>=0.9.27
dnspython>=2.4.0
psutil>=5.9.0